test: add inserts new tests
